Fix: 내체험관리 수정
